ms_distribucion/requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.5
django-filter==25.1
djangorestframework==3.16.1
drf-spectacular==0.28.0
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
mysqlclient==2.2.7
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
rpds-py==0.27.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
